+++ ENGINETIME.h
@@ -8,7 +8,7 @@ class EngineTime
 {
 public:
 	static void initialize();
-	static double getDeltaTime();
+	static float getDeltaTime();
 private:
 	EngineTime();
 	~EngineTime();
@@ -20,7 +20,7 @@ private:
 	std::chrono::system_clock::time_point start;
 	std::chrono::system_clock::time_point end;
 
-	double deltaTime = 0.0;
+	float deltaTime = 0.0;
 
 	static void LogFrameStart();
 	static void LogFrameEnd();
